Check when Malware Removal Tool was last updated
Added a function check_mrt_update which will check when MRT was last
updated by Apple. MRT functions as the other half of XProtect, wherein
it scans the macOS system on every boot for malware that is known to
Apple and then removes it automatically.

Similar to XProtect the returned version number could be compared to an
online source that monitors these. For example https://eclecticlight.co/