Output the log information to a file
Added a function output_to_file that will take an array of the output
information and store it in a .txt file. A function was made to do this
as the code will be repetitive.

At this stage, currently only has one function passing data to the
array. This is just a PoC to ensure that it works.

From here, the pretty colour formatting needs to be removed before the
logs are written to a file as they currently show as the text
representation of the colour code.

This is currently a preliminary implementation. The final product is
aimed to output to a .html or .PDF file.