Save logs to array and print them to terminal
Currently the next implementation from the previous commit. Next thing
to do is to create an 'errors' array so that the output to the file can
be easily split into "Passed" and "Errors" so that issues can be
highlighted easily.

On completion of the 'errors' array implementation, next step to save
to file.