Gather all data regarding Applications and signing
Added file applications.sh. Initially intended as a test however ended
up working alright. This script will gather a list of all the
Applications installed, where they were obtained from (Apple, Unknown
or Identified Developer) and the signing status.

The next step of this is to check notarization status of the signed
Applications. This may not be added depending on time constraints.