Copy data to a disk/random folder for collection
A new script has been started to perform the collection of the data.
This has been developed with the intention to allow for the data to be
copied to an external disk and then can be analyzed on a separate
device.

If no disk is passed to the script, it will create a randomly named
folder in ~/ (generated using /dev/urandom), collect the data in there
and then create an encrypted disk image to store said data in. This is
still to be implemented.