Verify that the disk exists and erase it.
This uses diskutil apfs to erase the disk passed to the script. It
encrypts it using a base64 passphrase from /dev/urandom and displays it
in the output.

An additional that was suggested was to allow for providing an IP
address and transmitting the data using nc or SSH. This would allow for
situations where a USB can't be used or isn't to hand.