Verify an entered IP address is valid
As the collected data can be transferred over the network, the user can
enter an IP address to send the data to. This IP address has to be
valid, in that it contains 4 '.' separated numbers, each containing
numbers between 0-9 and not being more than 255. This was checked using
a regular expression.

Next steps: compress collected data, generate a hash of the compressed
file and then netcat it across to the other device.