Create an encrypted disk image to store data
Using hdiutil, an APFS disk image is created and encrypted using
AES-128 with a randomly generated passpword. The size of the disk image
is calculated using 'du -sk', which allows for printing the size in a
human-readable format and in KB. An extra 100MB is added to the size of
the disk image as du seems to under-calculate the actual size (or at
least the size presented by Finder is different) which was causing the
disk image to be too small to copy the data.