Collect information about the Users.
Implementing cFiles function allows for gathering information about the
users on the device, other than those that are standard for macOS. This
is performed using 'find' to look for all files in each users directory
and then 'stat' each file to show the permissions and modified and
accessed times. The file is also hashed using SHA-256 and the name,
times, permissions and hash of each file is written to a separate file
for analysis.

A later stage may allow for uploading hashes of suspicious files to
VirusTotal, although this will be time dependent.