Collect both System and User Launch Agents.
For each user, the common directories for both Launch Agents and
Daemons are searched. These will typically hold Apple Launch Agents and
so these have been excluded from the 'find' command.

If Agents/Daemons exist, but aren't necessarily accompanied by a known-
good Application, then this can identify a piece of malware that is
persisting and launching on boot.