Separate the collection of User and System Agents.
A separate for loop has been implemented for collecting the User Agents.
This was previously implemented as the 'e' flag had to be flipped
temporarily. This was due to 'cp' giving an erroneous exit code.

This has since been fixed by replacing the 'cp' command with a 'find'
command. The for loop is still separate as it now allows for easier
collection of the Agents.