Parse and print the signing status of apps.
Added function print_signing(). Parses three seperate files, signed.txt,
 notsigned.txt and notarized.txt. These files all contain the
Applications that are signed, unsigned or signed and notarized. As
unsigned Applications are more suspicious, these are displayed in the
main .html file. These and the other Applications  are also added to a
another .html file so these can be checked if need be.