Analyse the collected data from browsers.
Added function analyse_browser(). This function will extract the
browsing and download  history from the .plist files and SQLite
databases extracted from the device.

To open the databases, these need to have r/w access, and so have to be
copied from the read-only disk image/USB. This is performed regardless
of the extraction medium as it allows for these to be deleted after
data extraction and a copy still to remain if further analysis is
required.