Analyse collected Launch Agents/Daemons.
Added function analyse_launch_agents(). If any exist, these will have
been extracted during collection. These will then be printed to the
main .html. At this time, it is quite difficult to detect if a
Launch Agent is malicious programmatically (at least for me), so this
may be added at a later stage.