Extract sudo commands from command history.
Reworked print_users() and added write_sudo(). If the .zsh/bash_history
files exist, then these will be scanned for any sudo commands that have
been executed. If there is a .*_history file for the `root` user, then
all of these commands are displayed, as the root user isn't really used
on macOS.

The remainder of the commands (and the sudo commands) are also added to
a second PDF so that all the commands can be viewed in order that they
were executed.