Improve the reliability of extracting Launch info.
Previously, whenever trying to extract the Launch Agents/Daemons, this
would fail. To fix this, set +e/-e had to be used so that this wouldn't
break on a non-zero status code.

Above that, it now just tries to cp the directory (doesn't really care
if it exists or not) as trying to check using if -d was having some
trouble.