Collect up-to-date zsh history.
The way that zsh operates, the history isn't written to a file line-by-
line. By checking if the .zsh_history file exists for each user, the
`history` command can be used to write an up-to-date version to a file.

Also, I'm an idiot and this build should now be passing as I added an
extra } at the end of a function in analysis.sh. Idiot.