Check signing status of Applications.
Reworked cApplication() to now add a chonk that allows for checking if
an Application is signed, unsigned (or has a broken signature, they're
the same in this respect) or signed and notarized. To check
Notarization, Xcode Tools must be installed, however, this will skip if
it isn't.

Also changed the format of how the data is displayed when executing the
`system_profiler` commands. It was a good idea in my head to have
'--command', BUT when it came to parsing it...Oh NELLY it was a pain.
This has been removed. Thank goodness.