Tidied extraction functions.
Added the call to collect() in usb() (because I'm an idiot and forgot
about it...).

Moved the hashing of files to before compression, because previously it
was after which was really stupid.

Altered the way that the data transfer is performed using netcat.
Previously it relied on calling tar on the current directory and then
piping that to nc. This worked, however, it failed if the listener
wasn't listening when it tried to connect. This has been fixed by
instead just passing the .tar archive that is created during normal
compression and adding '-w 30', which makes netcat wait 30 seconds for
a connection before closing it. This works a lot better than what was
implemented previously.