Use a main file to call anlaysis or collect.
Added a new file, macos-ir.sh, which will act as the initial
entry-point to the scripts. At this time it is empty, however, it is to
act as a replacement script which will accept arguments similar to
analysis.sh and collect.sh, and will also install the tools. This
doesn't need to be run with sudo, which fixes an issue that I was
having with installing the analysis tools in the original script (the
script needed to be run with sudo, but Homebrew won't install if
executed using sudo).