Correctly display the signed Applications.
In the function print_signing(), the leading '/' in file paths has been
removed. Reason for this is the directory that these files are saved in
is named 'Applications'. If the directory is written as '/Applications'
then this will use the macOS Applications folder, rather than the one
that I'm actually looking for.