Correctly Interpret -s Flag For Skipping Files.
Previously in collect.sh, the value of `SKIP` was changed to `false`
after it was passed from macos-ir.sh. This has been fixed.

The value in macos-ir.sh was also incorrectly collected, an extra
`shift` statement was added which caused an error as `OPTIND` would
shift passed the end. This has been fixed.