Correctly Perform Hash Of User Files in ~/Library.
Previously Users/[username]/Library/... was skipped when gathering the
users' file information. This was to reduce the run time of the script,
however, this has been removed as the Library/ directory can contain
a whole load of information if malware has been executed including
when new Launch Agents were created and executables that are hidden in
there.