Create Github Workflow for CI.
Created .github/workflows/lint.yml  to execute lint.sh on both push and pull request on branches master and wip. As lint.sh utilises Shellcheck, it has to be downloaded before it can be used. This action downloads the stable release of Shellcheck, ensures that it can be executed and execute lint.sh.