Check if Terminal.app has FDA.
Added function check_fda to test if Terminal has Full Disk Access. This
is performed using file against TCC.db to check if there is access.

TCC.db itself is protected under FDA and so if FDA is not granted, it
throws an error. If FDA is granted then it will continue as expected.

FDA is required in collect.sh as without it a lot of the data that is
intended (and planned in future) to be collected can't be.